requirements.txt
asgiref==3.8.1
#Django==5.1.3
python-dotenv==1.0.1
sqlparse==0.5.2
requests==2.31.0
pyTelegramBotAPI